# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deverdex agency website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  title: string;
  slug: string;
  description: string;
  category: string;
  platform: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  liveUrl?: string | null;
  tags: string[];
  featured: boolean;
  createdAt: string;
}

export interface ProjectInput {
  title: string;
  slug: string;
  description: string;
  category: string;
  platform: string;
  imageUrl?: string;
  liveUrl?: string;
  tags: string[];
  featured?: boolean;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  tags: string[];
  publishedAt: string;
  readingTime: number;
}

export interface BlogPostInput {
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  imageUrl?: string;
  tags: string[];
  readingTime: number;
}

export interface ServicePackage {
  name: string;
  price: string;
  delivery: string;
}

export interface Service {
  id: number;
  name: string;
  slug: string;
  description: string;
  features: string[];
  packages: ServicePackage[];
  icon: string;
  category: string;
}

export interface Experiment {
  id: number;
  title: string;
  slug: string;
  description: string;
  tags: string[];
  status: string;
  type: string;
}

export interface QuoteInput {
  name: string;
  email: string;
  company?: string;
  service: string;
  budget?: string;
  message: string;
}

export interface ContactInput {
  name: string;
  email: string;
  subject: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface SiteStats {
  projectsDelivered: number;
  happyClients: number;
  avgRating: number;
  countriesServed: number;
  yearsOfExperience: number;
}

export type GetProjectsParams = {
category?: string;
};