# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deverdex agency website API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Portfolio projects
  - name: blog
    description: Blog posts
  - name: services
    description: Services and pricing
  - name: experiments
    description: Experiments / R&D items
  - name: quote
    description: Quote requests
  - name: contact
    description: Contact messages
  - name: stats
    description: Site statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: List all portfolio projects
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

  /blog:
    get:
      operationId: getBlogPosts
      tags: [blog]
      summary: List blog posts
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found

  /services:
    get:
      operationId: getServices
      tags: [services]
      summary: List services with pricing
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /experiments:
    get:
      operationId: getExperiments
      tags: [experiments]
      summary: List experiments
      responses:
        "200":
          description: List of experiments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experiment"

  /quote:
    post:
      operationId: submitQuote
      tags: [quote]
      summary: Submit a quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Quote submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get site statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Project:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          type: string
        platform:
          type: string
        imageUrl:
          type: ["string", "null"]
        liveUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        createdAt:
          type: string
      required: [id, title, slug, description, category, platform, tags, featured, createdAt]

    ProjectInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          type: string
        platform:
          type: string
        imageUrl:
          type: string
        liveUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
      required: [title, slug, description, category, platform, tags]

    BlogPost:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        publishedAt:
          type: string
        readingTime:
          type: number
      required: [id, title, slug, excerpt, content, tags, publishedAt, readingTime]

    BlogPostInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        readingTime:
          type: number
      required: [title, slug, excerpt, content, tags, readingTime]

    Service:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        features:
          type: array
          items:
            type: string
        packages:
          type: array
          items:
            $ref: "#/components/schemas/ServicePackage"
        icon:
          type: string
        category:
          type: string
      required: [id, name, slug, description, features, packages, icon, category]

    ServicePackage:
      type: object
      properties:
        name:
          type: string
        price:
          type: string
        delivery:
          type: string
      required: [name, price, delivery]

    Experiment:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        type:
          type: string
      required: [id, title, slug, description, tags, status, type]

    QuoteInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        company:
          type: string
        service:
          type: string
        budget:
          type: string
        message:
          type: string
      required: [name, email, service, message]

    ContactInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
      required: [name, email, subject, message]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]

    SiteStats:
      type: object
      properties:
        projectsDelivered:
          type: number
        happyClients:
          type: number
        avgRating:
          type: number
        countriesServed:
          type: number
        yearsOfExperience:
          type: number
      required: [projectsDelivered, happyClients, avgRating, countriesServed, yearsOfExperience]