# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deverdex agency website API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  tags: string[];
  publishedAt: string;
  readingTime: number;
}