# lib/api-zod/src/generated/types/experiment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deverdex agency website API
 * OpenAPI spec version: 0.1.0
 */

export interface Experiment {
  id: number;
  title: string;
  slug: string;
  description: string;
  tags: string[];
  status: string;
  type: string;
}