# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deverdex agency website API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  title: string;
  slug: string;
  description: string;
  category: string;
  platform: string;
  imageUrl?: string;
  liveUrl?: string;
  tags: string[];
  featured?: boolean;
}