# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deverdex agency website API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteInput {
  name: string;
  email: string;
  company?: string;
  service: string;
  budget?: string;
  message: string;
}