# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deverdex agency website API
 * OpenAPI spec version: 0.1.0
 */
import type { ServicePackage } from './servicePackage';

export interface Service {
  id: number;
  name: string;
  slug: string;
  description: string;
  features: string[];
  packages: ServicePackage[];
  icon: string;
  category: string;
}