# lib/api-zod/src/generated/types/servicePackage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deverdex agency website API
 * OpenAPI spec version: 0.1.0
 */

export interface ServicePackage {
  name: string;
  price: string;
  delivery: string;
}