# lib/api-zod/src/generated/types/siteStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Deverdex agency website API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteStats {
  projectsDelivered: number;
  happyClients: number;
  avgRating: number;
  countriesServed: number;
  yearsOfExperience: number;
}